agents/voice/main.py
import asyncio
from pathlib import Path
from dotenv import load_dotenv
from livekit.agents import JobContext, WorkerOptions, cli, JobRequest
from livekit import rtc

load_dotenv(dotenv_path=Path(__file__).with_name(".env"))

# 1. 방에 참여했을 때 수행할 동작 (가공 공장의 실제 로직)
async def entrypoint(ctx: JobContext):
    print(f"--- 에이전트 방 입장 시도 중... 방 이름: {ctx.room.name} ---")

    await ctx.connect()
    # 접속 후 메타데이터를 설정해 프론트에서 필터링할 수 있게 합니다.
    await ctx.room.local_participant.set_metadata(
        '{"type": "agent", "name": "AI"}'
    )
    print(f"--- {ctx.room.name} 방에 접속 성공! 데이터 감시 시작 ---")

    # 참여자가 트랙(오디오/비디오)을 보낼 때 감지
    @ctx.room.on("track_subscribed")
    def on_track_subscribed(track: rtc.Track, publication: rtc.TrackPublication, participant: rtc.RemoteParticipant):
        if track.kind == rtc.TrackKind.KIND_AUDIO:
            print(f"[감지] 오디오 데이터 수신 시작: {participant.identity} (ID: {participant.sid})")
            # 나중에 여기에 AI(STT) 로직이 들어갑니다.

# 2. 서버의 일거리 요청(Job Request)을 어떻게 처리할지 결정
async def request_fnc(req: JobRequest):
    print(f"--- 새로운 일거리 요청 받음: {req.room.name} ---")
    # 모든 요청을 수락합니다. 
    # 실제 서비스에서는 여기서 특정 방만 수락하거나 권한을 확인할 수 있습니다.
    await req.accept(name="AI")

# 3. 워커 실행 설정
if __name__ == "__main__":
    print("--- AI 에이전트 워커 가동 시작 ---")
    cli.run_app(WorkerOptions(
        entrypoint_fnc=entrypoint,
        request_fnc=request_fnc, # ★ 이 부분을 추가해야 서버의 요청을 수락합니다.
    ))
