agents/requirements.txt
python-dotenv
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
